requirements.txt
fastapi==0.115.12
pydantic==2.11.1
PyJWT==2.10.1 
pyodbc==5.1.0
python-dotenv==1.0.0
uvicorn==0.29.0
